finish book api

+++ server/src/books/domain/book.usecases.ts
@@ -1,6 +1,6 @@
-import { Inject, Injectable } from "@nestjs/common";
+import { BadRequestException, Inject, Injectable } from "@nestjs/common";
 import { BookRepository, bookRepositoryKey } from "../domain/book.repository";
-import { BookCreateCommand, BookUpdateCommand } from "./book.model";
+import { BookCreateCommand, BookListResult, BookQueryParams, BookUpdateCommand, IBook } from "./book.model";
 
 
 @Injectable()
@@ -11,23 +11,31 @@ export class BookUseCase {
     private readonly repository: BookRepository
   ) { }
 
-  async getAllBooks() {
-    return this.repository.getAllBooks();
+  async getAllBooks(params: BookQueryParams): Promise<BookListResult> {
+    return await this.repository.getAllBooks(params);
   }
 
-  async getBookById(id: number) {
-    return this.repository.getBookById(id);
+  async getBookById(id: IBook["id"]) {
+    return await this.repository.getBookById(id);
   }
 
   async createBook(book: BookCreateCommand) {
-    return this.repository.createBook(book);
+    return await this.repository.createBook(book);
   }
 
   async updateBook(book: BookUpdateCommand) {
-    return this.repository.updateBook(book);
+    if (!await this.repository.getBookById(book.id)) {
+      throw new BadRequestException(`Invalid book id (${book.id})`);
+    }
+
+    return await this.repository.updateBook(book);
   }
 
-  async deleteBook(id: number) {
-    return this.repository.deleteBook(id);
+  async deleteBook(id: IBook["id"]) {
+    if (!await this.repository.getBookById(id)) {
+      throw new BadRequestException(`Invalid book id (${id})`);
+    }
+
+    return await this.repository.deleteBook(id);
   }
 }
